Shared descriptor sets

+++ Wulkan/src/engine/Frustum.cpp
@@ -30,7 +30,7 @@ void Frustum::set_camera_matrix(glm::mat4 proj_view_mat)
 {
 	glm::mat4 cam_inv_mat = glm::inverse(proj_view_mat);
 
-	for (int i = 0; i < 8; i++) {
+	for (size_t i = 0; i < 8; i++) {
 		glm::vec4 frustum_world_space = cam_inv_mat * glm::vec4(frustum_NDC[i], 1);
 		frustum_world_space /= frustum_world_space.w;
